% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\name{as.data.frame.ecip}
\alias{as.data.frame.ecip}
\title{\code{\link[=as.data.frame.ecip]{as.data.frame.ecip()}}}
\usage{
\method{as.data.frame}{ecip}(
  x,
  ...,
  row.title = sprintf(fmt = "\%s: \%s", x@endpoint, x@nobs)
)
}
\arguments{
\item{x}{\linkS4class{ecip} object}

\item{...}{..}

\item{row.title}{\link[base]{character} scalar}
}
\description{
\code{\link[=as.data.frame.ecip]{as.data.frame.ecip()}}
}
\keyword{internal}
