% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\name{as.matrix.ecip}
\alias{as.matrix.ecip}
\title{Convert \linkS4class{ecip} into \link[base]{matrix}}
\usage{
\method{as.matrix}{ecip}(x, fmt = "\%.2f", type = c("full", "ncol1", "p_only"), ...)
}
\arguments{
\item{x}{\linkS4class{ecip} object}

\item{fmt}{..}

\item{type}{\code{'full'} (default), \code{'ncol1'} or \code{p_only}}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
