% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC.R
\name{as.matrix.stepAIC}
\alias{as.matrix.stepAIC}
\title{Convert \link[MASS]{stepAIC} Return to \link[base]{matrix}}
\usage{
\method{as.matrix}{stepAIC}(x, ...)
}
\arguments{
\item{x}{returned object from function \link[MASS]{stepAIC}, or \link[MASS.tzh]{stepAIC_complete}.}

\item{...}{additional parameters, currently of no use}
}
\description{
S3 method dispatch for S3 generic \link[base]{as.matrix}.
}
\keyword{internal}
