% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{as.multivar}
\alias{as.multivar}
\alias{multivar}
\title{as.multivar}
\usage{
as.multivar(x, ...)
}
\arguments{
\item{x}{..}

\item{...}{parameters of \code{\link[=subset.univar]{subset.univar()}}}
}
\description{
as.multivar
}
\examples{
# see ?rmd_
}
