% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{as.univar}
\alias{as.univar}
\alias{univar}
\alias{as.univar.terms}
\alias{as.univar.formula}
\alias{as.univar.default}
\title{Convert into Multiple Univariable Structures}
\usage{
as.univar(object, ...)

\method{as.univar}{terms}(object, envir = parent.frame(), ...)

\method{as.univar}{formula}(formula, envir = parent.frame(), ...)

\method{as.univar}{default}(object, ...)
}
\arguments{
\item{object}{see \strong{Usage}}

\item{...}{additional parameters, currently of no}

\item{envir}{\link[base]{environment},
for \link[stats]{terms} and \link[stats]{formula} dispatches,
so that the returned univariable formula(s) are
in the same \link[base]{environment} as the input}

\item{formula}{\link[stats]{formula}}
}
\description{
Convert into Multiple Univariable Structures
}
\examples{
as.univar(y ~ x1 + I(x2^2))
as.univar(y ~ x1 + x2 + (1|g) + (1|g1:g2)) # e.g., ?ordinal::clmm
# see ?md_
}
\keyword{internal}
