% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eciplist.R
\name{as_flextable.eciplist}
\alias{as_flextable.eciplist}
\title{Convert an \code{eciplist} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{eciplist}(x, ...)
}
\arguments{
\item{x}{an \code{eciplist}, i.e., a \link[base]{list} of \linkS4class{ecip} objects}

\item{...}{potential parameters of the function \code{\link[=as_flextable.ecip]{as_flextable.ecip()}}}
}
\description{
Convert an \code{eciplist} to \link[flextable]{flextable}
}
\keyword{internal}
