% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{as_flextable.multivar}
\alias{as_flextable.multivar}
\title{as_flextable.multivar}
\usage{
\method{as_flextable}{multivar}(x, ...)
}
\arguments{
\item{x}{\link{multivar} object}

\item{...}{additional parameters, currently of no use}
}
\description{
..
}
