% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{as_flextable.univar}
\alias{as_flextable.univar}
\title{Convert \link{univar} Object to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{univar}(x, ...)
}
\arguments{
\item{x}{\link{univar} object}

\item{...}{additional parameters, currently of no use}
}
\value{
Function \code{\link[=as_flextable.univar]{as_flextable.univar()}} returns a \link[flextable]{flextable}.
}
\description{
..
}
