% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lm.R, R/confint.R
\name{confint_.mlm}
\alias{confint_.mlm}
\alias{confint_}
\alias{confint_.default}
\title{User-Related Confidence Interval}
\usage{
\method{confint_}{mlm}(x, level = 0.95, ...)

confint_(x, level, ...)

\method{confint_}{default}(x, level = 0.95, ...)
}
\arguments{
\item{x}{an R object}

\item{level}{\link[base]{double} scalar, default \code{.95}}

\item{...}{additional parameters for S3 dispatch}
}
\value{
Function \code{\link[=confint_]{confint_()}} returns
a \link[base]{ncol}-\code{2L} \link[base]{double} \link[base]{matrix},
with additional \link[base]{attributes},
\describe{
\item{\code{attr(.,'conf.level')}}{\link[base]{double} scalar, the argument \code{level}}
}
}
\description{
..
}
\keyword{internal}
