% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_.R
\name{desc_}
\alias{desc_}
\title{Model Description}
\usage{
desc_(x)
}
\arguments{
\item{x}{an R object}
}
\value{
The function \code{\link[=desc_]{desc_()}} returns a \link[base]{character} scalar.
}
\description{
..
}
\keyword{internal}
