% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_reg.R
\name{.md_reg}
\alias{.md_reg}
\title{Fast Markdown Lines for Regression Model}
\usage{
.md_reg(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=.md_reg]{.md_reg()}} returns an S4 class \code{'md_lines'}.
}
\description{
..
}
\note{
Old generic function \code{Sprintf()} and method \code{Sprintf.default()}
}
\keyword{internal}
