% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\docType{class}
\name{ecip}
\alias{ecip}
\alias{ecip-class}
\title{Coefficient Estimates, Confidence Intervals and \eqn{p}-Values}
\usage{
ecip(model)
}
\arguments{
\item{model}{see slot \verb{@model}}
}
\description{
..
}
\section{Slots}{

\describe{
\item{\code{model}}{original regression model}

\item{\code{coef}}{\link[base]{double} \link[base]{vector}}

\item{\code{p.value}}{\link[base]{double} \link[base]{vector}}

\item{\code{conf.int}}{\link[base]{ncol}-\code{2L} \link[base]{double} \link[base]{matrix}, with
an \link[base]{attributes} of \code{attr(.,'conf.level')},
inspired by the returned value of functions such as \link[stats]{t.test}.}

\item{\code{endpoint}}{..}

\item{\code{nobs}}{..}

\item{\code{exp}}{\link[base]{logical} scalar}

\item{\code{estnm}}{..}

\item{\code{note}}{..}
}}

