% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{endpoint}
\alias{endpoint}
\title{Endpoint}
\usage{
endpoint(x)
}
\arguments{
\item{x}{..}
}
\value{
The function \code{\link[=endpoint]{endpoint()}} returns a \link[base]{language} object.
}
\description{
Find the endpoint of an R regression model.
}
\examples{
endpoint(a ~ b)
# endpoint(a ~ b ~ c) # stop support
endpoint(a ~ b | c)

}
\keyword{internal}
