% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estnm.R
\name{expcoef}
\alias{expcoef}
\title{Taking Exponential of Coefficients}
\usage{
expcoef(x)
}
\arguments{
\item{x}{an R object}
}
\value{
The function \code{\link[=expcoef]{expcoef()}} returns a \link[base]{logical} scalar.
}
\description{
To determine whether to taking exponential of coefficients estimates and confidence intervals.
}
\keyword{internal}
