% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLink.R
\name{getCanonicalLink}
\alias{getCanonicalLink}
\title{Get Canonical Link Function of a Model Fit or \link[stats]{family}}
\usage{
getCanonicalLink(x)
}
\arguments{
\item{x}{an R object of model fit, or \link[stats]{family}}
}
\value{
The function \code{\link[=getCanonicalLink]{getCanonicalLink()}} returns a \link[base]{character} scalar.
}
\description{
..
}
\examples{
getCanonicalLink(binomial)
getCanonicalLink('binomial')
}
\keyword{internal}
