% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes_ext.R
\name{\%~\%}
\alias{\%~\%}
\title{Create Formula from Variable Names using only '+'}
\usage{
lhs \%~\% rhs
}
\arguments{
\item{lhs, rhs}{\link[base]{character} \link[base]{vector}s, variable names appearing in
the left- and right-hand-side of a \link[stats]{formula}}
}
\description{
..
}
\details{
Only for end-user!!!!

Function \link{\%~\%} appends two \link[base]{character} \link[base]{vector}s
of variable names into a \link[stats]{formula}.
This is much slower than \link[base]{~} operator, thus should only be used by end-user
}
\examples{
xs = c('age', 'sex')
character() \%~\% xs 
'.' \%~\% xs

.mapply(`\%~\%`, dots = list(rhs = c('age', 'sex')), MoreArgs = list(lhs = 'edp'))

}
