% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\name{intercept_rm}
\alias{intercept_rm}
\title{Remove Intercept Term(s)}
\usage{
intercept_rm(x)
}
\arguments{
\item{x}{\linkS4class{ecip} or \link[base]{matrix}}
}
\description{
Remove Intercept Term(s)
}
\keyword{internal}
