% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.stepCriterion.R
\name{md_.backwardCriterion}
\alias{md_.backwardCriterion}
\title{md_.backwardCriterion}
\usage{
\method{md_}{backwardCriterion}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
md_.backwardCriterion
}
