% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listof.R
\name{md_.listof}
\alias{md_.listof}
\title{\pkg{rmarkdown} Lines based on \link[flextable]{flextable}}
\usage{
\method{md_}{listof}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
\pkg{rmarkdown} Lines based on \link[flextable]{flextable}
}
\examples{
# none yet..

}
\keyword{internal}
