% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.stepCriterion.R
\name{md_.stepCriterion}
\alias{md_.stepCriterion}
\title{md_.stepCriterion}
\usage{
\method{md_}{stepCriterion}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
md_.stepCriterion
}
