% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_autoplot.R
\name{md_autoplot_}
\alias{md_autoplot_}
\title{R Markdown Lines by \code{ggplot2::autoplot}}
\usage{
md_autoplot_(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines by \code{ggplot2::autoplot}
}
\keyword{internal}
