% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ecips.R
\name{md_ecips}
\alias{md_ecips}
\title{R Markdown Lines for via Multiple \link[ecip]{ecip}s}
\usage{
md_ecips(x, xnm, ...)
}
\arguments{
\item{x}{an R object convertible to multiple \link[ecip]{ecip}s}

\item{xnm}{\link[base]{character} scalar}

\item{...}{..}
}
\description{
R Markdown Lines for via Multiple \link[ecip]{ecip}s
}
\details{
\itemize{
\item{\code{'mlm'} object, returned object from multiple endpoints in \link[stats]{lm};}
\item{\code{?pscl::zeroinfl}}
}
}
\keyword{internal}
