% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_flextable.R
\name{md_flextable_}
\alias{md_flextable_}
\title{R Markdown by \link[flextable]{flextable}}
\usage{
md_flextable_(x, xnm, font.size, ...)
}
\arguments{
\item{x, xnm, font.size, ...}{..}
}
\description{
R Markdown by \link[flextable]{flextable}
}
\keyword{internal}
