% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ecip.R, R/md_ecips.R
\name{md_lm}
\alias{md_lm}
\alias{md_.lm}
\alias{md_.mlm}
\title{R Markdown Lines for \link[stats]{lm} Model}
\usage{
\method{md_}{lm}(x, xnm, ...)

\method{md_}{mlm}(x, xnm, ...)
}
\arguments{
\item{x}{a regression model returned from function \link[stats]{lm} or \link[stats]{glm}}

\item{xnm}{\link[base]{character} scalar}

\item{...}{..}
}
\description{
R Markdown Lines for \link[stats]{lm} Model
}
\examples{
list(
 '`lm`' = lm(Sepal.Length ~ Species, data = iris)
) |> rmd.tzh::render_(file = 'lm')

list(
  '`mlm`' = lm(cbind(Sepal.Length, Petal.Length) ~ Species, data = iris)
) |> rmd.tzh::render_(file = 'mlm')
}
\keyword{internal}
