% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_multiple.R
\name{md_multiple}
\alias{md_multiple}
\alias{md_.mlm}
\alias{md_.zeroinfl}
\title{\pkg{rmarkdown} Lines of Multiple Models}
\usage{
\method{md_}{mlm}(x, xnm, ...)

\method{md_}{zeroinfl}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
\pkg{rmarkdown} Lines of Multiple Models
}
\examples{
library(rmd.tzh)

list(
  '`mlm`' = lm(cbind(Sepal.Length, Petal.Length) ~ Species, data = iris)
) |> render_(file = 'mlm')
library(pscl.tzh); list(
 '`zeroinfl`' = zeroinfl(art ~ . | 1, data = bioChemists)
) |> render_(file = 'zeroinfl')
}
\keyword{internal}
