% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_multiple.R
\name{md_multiple_}
\alias{md_multiple_}
\title{R Markdown Lines for Multiple Models}
\usage{
md_multiple_(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
R Markdown Lines for Multiple Models
}
