% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_regression.R
\name{md_regression_}
\alias{md_regression_}
\title{Fast Markdown Lines for Regression Model}
\usage{
md_regression_(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=md_regression_]{md_regression_()}} returns an S4 class \code{'md_lines'}.
}
\description{
..
}
\keyword{internal}
