% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lm.R, R/pval.R
\name{.pval.summary.lm}
\alias{.pval.summary.lm}
\alias{.pval.summary.glm}
\alias{pval}
\alias{.pval}
\alias{.pval.default}
\alias{.pval.coeftest}
\alias{.pval.listof}
\title{Retrieve \eqn{p}-values from Regression Model}
\usage{
\method{.pval}{summary.lm}(x)

\method{.pval}{summary.glm}(x)

.pval(x)

\method{.pval}{default}(x)

\method{.pval}{coeftest}(x)

\method{.pval}{listof}(x)
}
\arguments{
\item{x}{an R regression model}
}
\description{
..
}
\details{
Function \code{\link[=.pval.listof]{.pval.listof()}} applies to \code{'mlm' |> stats:::summary.mlm() |> .pval.listof()}.
}
\keyword{internal}
