% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineTerm.R
\name{refineTerm}
\alias{refineTerm}
\alias{refineTerm_multinom}
\alias{refineTerm_logical}
\alias{refineTerm_factor}
\alias{rename_xlevels}
\title{refineTerm}
\usage{
refineTerm(x, ...)

refineTerm_multinom(x, model, ...)

refineTerm_logical(x, model, ...)

refineTerm_factor(x, model, ...)

rename_xlevels(model)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}

\item{...}{..}

\item{model}{..}
}
\description{
..
}
\examples{
m1 = lm(mpg ~ cyl + am, data = mtc)
refineTerm(names(coef(m1)), model = m1)
summary(m2 <- lm(breaks ~ wool + tension, data = warpbreaks))
summary(m2a <- lm(breaks ~ -1 + wool + tension, data = warpbreaks))
rename_xlevels(m2)
rename_xlevels(m2a) # actually this is not ideal (but not wrong anyway) ..
refineTerm_factor(names(coef(m2)), model = m2)
}
\keyword{internal}
