% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.R, R/rmd_autoplot.R, R/rmd_flextable.R,
%   R/rmd_multiple.R
\name{rmd_}
\alias{rmd_}
\alias{rmd_.default}
\alias{rmd_.survival_roc}
\alias{rmd_.lcmm}
\alias{rmd_.listof}
\alias{rmd_.multivar}
\alias{rmd_.stepAIC}
\alias{rmd_.aov}
\alias{rmd_.TukeyHSD}
\alias{rmd_.glht}
\alias{rmd_.binTab}
\alias{rmd_.DemographicTable}
\alias{rmd_.summary.matchit}
\alias{rmd_.matchit}
\alias{rmd_.mlm}
\alias{rmd_.zeroinfl}
\title{Writing R Objects to \pkg{rmarkdown} Document}
\usage{
\method{rmd_}{default}(x, xnm, ...)

\method{rmd_}{survival_roc}(x, xnm, ...)

\method{rmd_}{lcmm}(x, xnm, ...)

\method{rmd_}{listof}(x, xnm, font.size, ...)

\method{rmd_}{multivar}(x, xnm, font.size, ...)

\method{rmd_}{stepAIC}(x, xnm, font.size, ...)

\method{rmd_}{aov}(x, xnm, font.size, ...)

\method{rmd_}{TukeyHSD}(x, xnm, font.size, ...)

\method{rmd_}{glht}(x, xnm, ...)

\method{rmd_}{binTab}(x, xnm, ...)

\method{rmd_}{DemographicTable}(x, xnm, font.size = 9, ...)

\method{rmd_}{summary.matchit}(x, xnm, font.size, ...)

\method{rmd_}{matchit}(x, xnm, ...)

\method{rmd_}{mlm}(x, xnm, ...)

\method{rmd_}{zeroinfl}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}

\item{font.size}{see function \link[flextable]{set_flextable_defaults}}
}
\description{
Writing R Objects to \pkg{rmarkdown} Document
}
\examples{
list(
 '`lm`' = lm(Sepal.Length ~ Species, data = iris)
) |> render_(file = 'lm')

list(
 'one `lme`' = ortho1,
 'another `lme`' = ortho2,
 '`gls`' = ovary1
) |> render_(file = 'lme_gls')

library(geepack); list(
 '`geeglm`' = geeglm(breaks ~ tension, id = wool, data = warpbreaks, corstr = 'exchangeable')
) |> render_(file = 'geeglm')

library(multcomp); list(
 '`glht` via `aov`' = aov(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett')),
 '`glht` via `lm`, single `$focus`' = lm(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey')),
 '`glht` via `lm`, multiple `$focus`' = lm(breaks ~ tension + wool, data = warpbreaks) |> 
   glht(linfct = mcp(tension = 'Tukey', wool = 'Dunnett'))
) |> render_(file = 'glht')

library(ordinal); list(
 '`clm`' = clm(rating ~ temp + contact, data = wine), 
 '`clmm`' = clmm(rating ~ temp + contact + (1|judge), data = wine)
) |> render_(file = 'clm_clmm')

list(
 '`coxph`' = coxph(os ~ age + ph.ecog, data = lung2)
) |> render_(file = 'coxph')

library(lme4); list(
 '`glmerMod`' = glmer(cbind(incidence, size-incidence) ~ period + (1|herd), 
     data = cbpp, family = binomial)
) |> render_(file = 'glmerMod')
  
library(ltm); list(
 '`cronbachAlpha`' = cronbach.alpha(LSAT, CI = TRUE, B = 500)
) |> render_(file = 'cronbachAlpha')

library(MASS); list(
 '`rlm`' = rlm(stack.loss ~ ., data = stackloss)
) |> render_(file = 'rlm')

library(VGAM)
pneumo = transform(pneumo, let = log(exposure.time))
list(
 '`vglm`' = vglm(cbind(normal, mild, severe) ~ let, propodds, data = pneumo)
) |> render_(file = 'vglm')

library(vcd); list(
 '`Kappa`' = Kappa(SexualFun)
) |> render_(file = 'Kappa')
data(mayo, package = 'survivalROC'); list(
 'survival_roc' = mayo |> 
  within.data.frame(expr = {
   time = as.difftime(time, units = 'days')
   edp = Surv(time, censor)
  }) |>
  survival_roc(formula = edp ~ mayoscore4, predict.time = 365)
) |> render_(file = 'survival_roc')
library(lcmm.tzh); list(
 '`lcmm`' = m20
) |> render_(file = 'lcmm')
list(
  multivar = lm(mpg ~ cyl + am + hp + wt + qsec + drat + disp, data = mtc) |>
     as.univar() |>
     as.multivar(subset = min_pvalue < .1)
) |> render_(file = 'multivar')
list(
  '`TukeyHSD`' = aov(breaks ~ wool + tension, data = warpbreaks) |> 
    TukeyHSD(which = 'tension', ordered = TRUE)
) |> render_(file = 'aov_etc')
library(DanielBiostatistics10th); list(
  '`binTab`' = binTab(array(c(7L, 3L, 8L, 6L), dim = c(2,2)))
) |> render_(file = 'binTab')
library(DemographicTable); list(
  '`DemographicTable`' = DemographicTable(CO2, groups = 'Type', include = c('conc', 'uptake'))
) |> render_(file = 'DemographicTable')
library(MatchIt)
m = matchit(treat ~ age+educ+race+nodegree+married+re74+re75, data = lalonde)
list(
 '`matchit1`' = m,
 '`matchit2`' = m |> summary(addlvariables = 're78')
) |> render_(file = 'matchit')
list(
  '`mlm`' = lm(cbind(Sepal.Length, Petal.Length) ~ Species, data = iris)
) |> render_(file = 'mlm')
library(pscl); list(
 '`zeroinfl`' = zeroinfl(art ~ . | 1, data = bioChemists)
) |> render_(file = 'zeroinfl')
}
