% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.R
\name{Sprintf}
\alias{Sprintf}
\alias{Sprintf.default}
\title{String Formatting of Statistical Models}
\usage{
Sprintf(x)

\method{Sprintf}{default}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=Sprintf]{Sprintf()}} returns a \link[base]{noquote} \link[base]{character} scalar
}
\description{
..
}
\keyword{internal}
