% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.stepAIC.R
\name{stepAIC}
\alias{stepAIC}
\alias{as_flextable.stepAIC}
\alias{as.matrix.stepAIC}
\title{Turn \link[MASS.tzh]{stepAIC_complete} Return to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{stepAIC}(
  x,
  row.title = attr(z, which = "row.title", exact = TRUE),
  hline_i = attr(z, which = "hline_i", exact = TRUE),
  vline_j = attr(z, which = "vline_j", exact = TRUE),
  ...
)

\method{as.matrix}{stepAIC}(x, ...)
}
\arguments{
\item{x}{returned value of function \link[MASS.tzh]{stepAIC_complete}}

\item{row.title, hline_i, vline_j}{..}

\item{...}{..}
}
\description{
Turn \link[MASS.tzh]{stepAIC_complete} Return to \link[flextable]{flextable}
}
\keyword{internal}
