% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\name{[.ecip}
\alias{[.ecip}
\title{Subset \linkS4class{ecip} Object}
\usage{
\method{[}{ecip}(x, i, ...)
}
\arguments{
\item{x}{an \linkS4class{ecip} object}

\item{i}{..}

\item{...}{..}
}
\description{
Subset \linkS4class{ecip} Object
}
\examples{
(x = lm(breaks ~ tension + wool, data = warpbreaks) |> ecip())
x[2:3]

}
\keyword{internal}
