% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar.R
\name{subset.univar}
\alias{subset.univar}
\title{subset.univar}
\usage{
\method{subset}{univar}(x, subset, ...)
}
\arguments{
\item{x}{\link{as.univar} object}

\item{subset}{..}

\item{...}{..}
}
\description{
subset.univar
}
