% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vterms.R
\name{vterms}
\alias{vterms}
\title{Variables in Terms}
\usage{
vterms(x)
}
\arguments{
\item{x}{an \code{R} model fit}
}
\description{
Variables in Terms
}
\examples{
# see ?ordinal::clmm examples
(SURENESS ~ PROD + (1|RESP) + (1|RESP:PROD)) |>
  vterms()
}
\keyword{internal}
