% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlevels.R
\name{xlevels}
\alias{xlevels}
\alias{xlevels.default}
\title{Wrapper for \link[stats]{.getXlevels}}
\usage{
xlevels(x)

\method{xlevels}{default}(x)
}
\arguments{
\item{x}{an R object}
}
\description{
..
}
\details{
Function \code{\link[=xlevels]{xlevels()}} is a wrapper of function \link[stats]{.getXlevels}.
}
\note{
Function \link[stats]{.getXlevels} only reports \link[base]{factor} levels,
\link[base]{logical} predictors will \emph{not} be included!
}
