% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\name{as.multivar}
\alias{as.multivar}
\alias{multivar}
\title{as.multivar}
\usage{
as.multivar(x, ...)
}
\arguments{
\item{x}{..}

\item{...}{parameters of \code{\link[=subset.univar]{subset.univar()}}}
}
\description{
as.multivar
}
\examples{
lm(mpg ~ cyl + am + vs + hp + wt + qsec + drat + disp, data = mtc) |>
 as.univar() |>
 as.multivar(subset = min_pvalue < .1)
}
