% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\name{as_flextable.ecip}
\alias{as_flextable.ecip}
\title{Convert \linkS4class{ecip} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{ecip}(x, ...)
}
\arguments{
\item{x}{\linkS4class{ecip} object}

\item{...}{potential parameters of function \code{\link[=as.data.frame.ecip]{as.data.frame.ecip()}}}
}
\description{
..
}
\keyword{internal}
