% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\name{as_flextable.ecip}
\alias{as_flextable.ecip}
\title{Convert \linkS4class{ecip} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{ecip}(x, row.title = sprintf(fmt = "\%s: \%s", x@endpoint, x@nobs), ...)
}
\arguments{
\item{x}{\linkS4class{ecip} object}

\item{row.title}{..}

\item{...}{potential parameters of functions \code{\link[=as.matrix.ecip]{as.matrix.ecip()}}
and \link[flextable.tzh]{as_flextable.matrix}}
}
\description{
..
}
