% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listof.R
\name{as_flextable.listof}
\alias{as_flextable.listof}
\title{Convert \code{'listof'} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{listof}(x, ...)
}
\arguments{
\item{x}{\code{'listof'}}

\item{...}{..}
}
\description{
Convert \code{'listof'} to \link[flextable]{flextable}
}
\note{
This is quite exploratory
}
