% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lm.R, R/coef.R
\name{coef_.mlm}
\alias{coef_.mlm}
\alias{coef_}
\alias{coef_.default}
\title{User-Related Coefficient Estimates}
\usage{
\method{coef_}{mlm}(x)

coef_(x)

\method{coef_}{default}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=coef_]{coef_()}} returns either a named \link[base]{numeric} \link[base]{vector},
or a \link[base]{list} of named \link[base]{numeric} \link[base]{vector}s.
}
\description{
User-Related Coefficient Estimates
}
\keyword{internal}
