% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataClasses.R
\name{dataClasses}
\alias{dataClasses}
\alias{dataClasses.default}
\alias{dataClasses.terms}
\title{dataClasses}
\usage{
dataClasses(x)

\method{dataClasses}{default}(x)

\method{dataClasses}{terms}(x)
}
\arguments{
\item{x}{a regression model}
}
\value{
S3 generic function \code{\link[=dataClasses]{dataClasses()}} returns a named \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\details{
Function \code{\link[=dataClasses]{dataClasses()}} is inspired by \code{attr(, 'dataClasses')} of
\link[stats]{terms} object.
}
\examples{
x = lm(breaks ~ tension + wool, data = warpbreaks)
x |> terms() |> attr(which = 'dataClasses')
dataClasses(x)

}
\keyword{internal}
