% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lm.R, R/desc_.R
\name{desc_.lm}
\alias{desc_.lm}
\alias{desc_.glm}
\alias{desc_.mlm}
\alias{desc}
\alias{desc_}
\alias{desc_.default}
\alias{desc_.family}
\title{Model Description}
\usage{
\method{desc_}{lm}(x)

\method{desc_}{glm}(x)

\method{desc_}{mlm}(x)

desc_(x)

\method{desc_}{default}(x)

\method{desc_}{family}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=desc_]{desc_()}} returns a \link[base]{character} scalar.
}
\description{
..
}
\keyword{internal}
