% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{endpoint}
\alias{endpoint}
\alias{endpoint.default}
\alias{endpoint.formula}
\title{Endpoint}
\usage{
endpoint(x)

\method{endpoint}{default}(x)

\method{endpoint}{formula}(formula)
}
\arguments{
\item{x}{..}

\item{formula}{\link[stats]{formula}}
}
\value{
Function \code{\link[=endpoint]{endpoint()}} returns a \link[base]{language} object.
}
\description{
Find the endpoint of an R regression model.
}
\examples{
endpoint(a ~ b)
# endpoint(a ~ b ~ c) # stop support
endpoint(a ~ b | c)
}
