% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estnm.R
\name{estnm}
\alias{estnm}
\alias{estnm.default}
\alias{estnm.glht}
\title{Proper Name of Estimates}
\usage{
estnm(x)

\method{estnm}{default}(x)

\method{estnm}{glht}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=estnm]{estnm()}} returns a \link[base]{character} scalar.
}
\description{
Proper name of estimates for various regression models
}
\keyword{internal}
