% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLink.R
\name{getCanonicalLink}
\alias{getCanonicalLink}
\alias{getCanonicalLink.default}
\alias{getCanonicalLink.family}
\alias{getCanonicalLink.function}
\alias{getCanonicalLink.character}
\title{getCanonicalLink}
\usage{
getCanonicalLink(x)

\method{getCanonicalLink}{default}(x)

\method{getCanonicalLink}{family}(x)

\method{getCanonicalLink}{`function`}(x)

\method{getCanonicalLink}{character}(x)
}
\arguments{
\item{x}{..}
}
\description{
..
}
\examples{
getCanonicalLink(binomial)
getCanonicalLink('binomial')
}
