% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLink.R
\name{getLink}
\alias{getLink}
\alias{getLink.family}
\alias{getLink.default}
\title{Get Link Function of a Model Fit}
\usage{
getLink(x)

\method{getLink}{family}(x)

\method{getLink}{default}(x)
}
\arguments{
\item{x}{see \strong{Usage}}
}
\description{
Get link function of a model fit,
and determine whether the model is using the canonical link of the family!
}
\details{
Not sure how useful \code{marked:::inverse.link} will be in future!
}
\keyword{internal}
