% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecip.R
\name{intercept_rm}
\alias{intercept_rm}
\alias{intercept_rm.ecip}
\alias{intercept_rm.matrix}
\title{\link{intercept_rm}}
\usage{
intercept_rm(x)

\method{intercept_rm}{ecip}(x)

\method{intercept_rm}{matrix}(x)
}
\arguments{
\item{x}{..}
}
\description{
\link{intercept_rm}
}
\keyword{intercept_rm}
