% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.R, R/md_autoplot.R, R/md_multiple.R
\name{md_}
\alias{md_}
\alias{md_.lm}
\alias{md_.lme}
\alias{md_.gls}
\alias{md_.glmgee}
\alias{md_.clm}
\alias{md_.clmm}
\alias{md_.coxph}
\alias{md_.merMod}
\alias{md_.rlm}
\alias{md_.vlm}
\alias{md_.lcmm}
\alias{md_.mlm}
\alias{md_.zeroinfl}
\title{\pkg{rmarkdown} Lines based on \code{\link[=ecip]{ecip()}}}
\usage{
\method{md_}{lm}(x, xnm, ...)

\method{md_}{lme}(x, xnm, ...)

\method{md_}{gls}(x, xnm, ...)

\method{md_}{glmgee}(x, xnm, ...)

\method{md_}{clm}(x, xnm, ...)

\method{md_}{clmm}(x, xnm, ...)

\method{md_}{coxph}(x, xnm, ...)

\method{md_}{merMod}(x, xnm, ...)

\method{md_}{rlm}(x, xnm, ...)

\method{md_}{vlm}(x, xnm, ...)

\method{md_}{lcmm}(x, xnm, ...)

\method{md_}{mlm}(x, xnm, ...)

\method{md_}{zeroinfl}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
\pkg{rmarkdown} Lines based on \code{\link[=ecip]{ecip()}}
}
\examples{
library(rmd.tzh)

list(
 '`lm`' = lm(Sepal.Length ~ Species, data = iris)
) |> render_(file = 'lm')

library(nlme); list(
 '`lme`' = lme(distance ~ age, data = Orthodont, keep.data = TRUE),
 '`gls`' = gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), 
     data = Ovary, correlation = corAR1(form = ~ 1 | Mare))
) |> render_(file = 'lme_gls')

library(glmtoolbox); list(
 '`glmgee`' = glmgee(breaks ~ tension, id = wool, data = warpbreaks, corstr = 'exchangeable')
) |> render_(file = 'glmgee')

library(ordinal); list(
 '`clm`' = clm(rating ~ temp + contact, data = wine), 
 '`clmm`' = clmm(rating ~ temp + contact + (1|judge), data = wine)
) |> render_(file = 'clm_clmm')

library(lme4); list(
 '`glmerMod`' = glmer(cbind(incidence, size-incidence) ~ period + (1|herd), 
     data = cbpp, family = binomial)
) |> render_(file = 'glmerMod')
  
library(MASS); list(
 '`rlm`' = rlm(stack.loss ~ ., data = stackloss)
) |> render_(file = 'rlm')

library(VGAM)
pneumo = transform(pneumo, let = log(exposure.time))
list(
 '`vglm`' = vglm(cbind(normal, mild, severe) ~ let, propodds, data = pneumo)
) |> render_(file = 'vglm')

library(lcmm.tzh); list(
 '`lcmm`' = m20
) |> render_(file = 'lcmm')
list(
  '`mlm`' = lm(cbind(Sepal.Length, Petal.Length) ~ Species, data = iris)
) |> render_(file = 'mlm')
library(pscl.tzh); list(
 '`zeroinfl`' = zeroinfl(art ~ . | 1, data = bioChemists)
) |> render_(file = 'zeroinfl')
}
