% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.R, R/md_multiple.R
\name{md_}
\alias{md_}
\alias{md_.lm}
\alias{md_.rlm}
\alias{md_.mlm}
\title{\pkg{rmarkdown} Lines based on \code{\link[=ecip]{ecip()}}}
\usage{
\method{md_}{lm}(x, xnm, ...)

\method{md_}{rlm}(x, xnm, ...)

\method{md_}{mlm}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
\pkg{rmarkdown} Lines based on \code{\link[=ecip]{ecip()}}
}
\examples{
library(rmd.tzh)

list(
 '`lm`' = lm(Sepal.Length ~ Species, data = iris)
) |> render_(file = 'lm')

library(MASS); list(
 '`rlm`' = rlm(stack.loss ~ ., data = stackloss)
) |> render_(file = 'rlm')

list(
  '`mlm`' = lm(cbind(Sepal.Length, Petal.Length) ~ Species, data = iris)
) |> render_(file = 'mlm')
}
