% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ecip.R
\name{md_ecip}
\alias{md_ecip}
\title{R Markdown Lines via \link[ecip]{ecip}}
\usage{
md_ecip(x, xnm, ...)
}
\arguments{
\item{x}{an R object convertible to one \link[ecip]{ecip}}

\item{xnm, ...}{..}
}
\description{
R Markdown Lines via \link[ecip]{ecip}
}
\keyword{internal}
