% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_flextable.R
\name{md_flextable}
\alias{md_flextable}
\alias{md_.listof}
\title{\pkg{rmarkdown} Lines based on \link[flextable]{flextable}}
\usage{
\method{md_}{listof}(x, xnm, font.size, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{font.size}{see function \link[flextable]{set_flextable_defaults}}
}
\description{
\pkg{rmarkdown} Lines based on \link[flextable]{flextable}
}
\examples{
# library(rmd.tzh)
# none yet..

}
\keyword{internal}
