% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\alias{.pval}
\alias{.pval.default}
\alias{.pval.listof}
\title{Retrieve \eqn{p}-values from Regression Model}
\usage{
.pval(x)

\method{.pval}{default}(x)

\method{.pval}{listof}(x)
}
\arguments{
\item{x}{an R regression model}
}
\description{
..
}
