% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.R, R/aov.R, R/listof.R
\name{Sprintf}
\alias{Sprintf}
\alias{Sprintf.default}
\alias{Sprintf.aov}
\alias{Sprintf.TukeyHSD}
\alias{Sprintf.listof}
\alias{Sprintf.stepAIC}
\title{String Formatting of Statistical Models}
\usage{
Sprintf(x)

\method{Sprintf}{default}(x)

\method{Sprintf}{aov}(x)

\method{Sprintf}{TukeyHSD}(x)

\method{Sprintf}{listof}(x)

\method{Sprintf}{stepAIC}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \link{Sprintf} returns a \link[base]{noquote} \link[base]{character} scalar
}
\description{
..
}
