% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.R, R/listof.R, R/stepAIC.R,
%   R/stepCriterion.R
\name{Sprintf}
\alias{Sprintf}
\alias{Sprintf.default}
\alias{Sprintf.listof}
\alias{Sprintf.stepAIC}
\alias{Sprintf.backwardCriterion}
\title{String Formatting of Statistical Models}
\usage{
Sprintf(x)

\method{Sprintf}{default}(x)

\method{Sprintf}{listof}(x)

\method{Sprintf}{stepAIC}(x)

\method{Sprintf}{backwardCriterion}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=Sprintf]{Sprintf()}} returns a \link[base]{noquote} \link[base]{character} scalar
}
\description{
..
}
\keyword{internal}
