% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vterms.R
\name{vterms}
\alias{vterms}
\alias{vterms.default}
\title{Variables in Terms}
\usage{
vterms(x)

\method{vterms}{default}(x)
}
\arguments{
\item{x}{an \code{R} model fit}
}
\description{
Variables in Terms
}
\examples{
library(ordinal)
m = clmm(SURENESS ~ PROD + (1|RESP) + (1|RESP:PROD), data = soup,
 link = 'probit', threshold = 'equidistant')
m |> vterms()
}
\keyword{internal}
